dash-chatbot-app/requirements.txt
dash==3.0.2
dash-bootstrap-components==2.0.0
mlflow>=2.21.2
python-dotenv==1.1.0
databricks-sdk
reportlab>=4.0.0
